test(lib-nexus): drop unnecesary config
